requirements.txt
# Core dependencies
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Data processing
scikit-learn>=1.3.0
pyarrow>=12.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Training utilities
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0

# Utilities
pyyaml>=6.0
omegaconf>=2.3.0
hydra-core>=1.3.0

# Testing
pytest>=7.3.0
pytest-cov>=4.1.0

# Development
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0
